+++ NEWS.md
@@ -1,5 +1,6 @@
-# TreeTools 1.16.1.9000 (development) #
+# TreeTools 1.16.1.9001` (development) #
 
+- New method `as.Splits.integer()`.
 - Fix `RoguePlot(sort = TRUE)`
   [Rogue#33](https://github.com/ms609/Rogue/issues/33).